feat: support for Symfony Serializer 6+
