+++ tests/Fixtures/TestBundle/Entity/TypedScalarValueTrait.php
@@ -1,0 +1,32 @@
+<?php
+
+/*
+ * (c) Kévin Dunglas <dunglas@gmail.com>
+ *
+ * This source file is subject to the MIT license that is bundled
+ * with this source code in the file LICENSE.
+ */
+
+namespace Dunglas\DoctrineJsonOdm\Tests\Fixtures\TestBundle\Entity;
+
+if (PHP_MAJOR_VERSION < 8) {
+    return;
+}
+
+use Symfony\Component\Serializer\Normalizer\DenormalizerInterface;
+use Symfony\Component\Serializer\Normalizer\NormalizerInterface;
+
+trait TypedScalarValueTrait
+{
+    public function normalize(NormalizerInterface $normalizer, string $format = null, array $context = []): array|string|int|float|bool
+    {
+        return $this->value;
+    }
+
+    public function denormalize(DenormalizerInterface $denormalizer, array|string|int|float|bool $data, string $format = null, array $context = [])
+    {
+        $this->value = $data;
+
+        return $this;
+    }
+}
